GeometricObjects

Installation

pkg> add https://github.com/KeitaNakamura/GeometricObjects.jl.git

Types and Functions

GeometricObjects.Line — Type

Line(a::Vec, b::Vec)
Line(a::Vec => b::Vec)

Base.in — Method

in(x::Vec, line::Line)

Check if x is in line.

Examples

julia> line = Line(@Vec[0.0, 0.0], @Vec[2.0, 2.0])
Line{2, Float64}:
  Coordinates: [[0.0, 0.0], [2.0, 2.0]]
  Attitude: [1.0, 0.0]

julia> @Vec[1.0, 1.0] in line
true

julia> @Vec[1.0, 0.0] in line
false

Base.in — Method

in(x::Vec, ::Polygon; include_bounds = true)

Check if x is in a polygon.

Base.in — Method

in(x::Vec, ::Sphere; include_bounds = true)

Check if x is in a sphere.

julia> sphere = Sphere(Vec(1.0,1.0), 1.0)
Sphere{2, Float64}:
  Coordinates: [[1.0, 1.0]]
  Attitude: [1.0, 0.0]

julia> Vec(0.5, 0.5) in sphere
true

julia> Vec(0.0, 0.0) in sphere
false

Base.intersect — Method

intersect(::Line, ::Line; [extended = (false, false)])

Find intersection point from two lines. Return nothing if not found.

Base.intersect — Method

intersect(::Polygon, ::Line; [extended = false])

Find the closest intersection point from line to polygon. Return nothing if not found.

Base.intersect — Method

intersect(::Polyline, ::Line; [extended = false])

Find the closest intersection point from line to polyline. Return nothing if not found.

GeometricObjects.distance — Method

distance(::Line, x::Vec)
distance(::Line, x::Vec, threshold::Real)

Compute the distance vector from x to perpendicular foot. When threshold is given, check the contact between line and point x, and return nothing if contact is not detected. Note that if the perpendicular foot does not lie on the line, contact detection is performed using distance between x and vertices of line.

julia> line = Line(@Vec[0.0, 0.0] => @Vec[1.0, 1.0])
Line{2, Float64}:
  Coordinates: [[0.0, 0.0], [1.0, 1.0]]
  Attitude: [1.0, 0.0]

julia> distance(line, @Vec[1.0, 0.0])
2-element Vec{2, Float64}:
 -0.5
  0.5

julia> distance(line, @Vec[1.0, 0.0], 1.0)
2-element Vec{2, Float64}:
 -0.5
  0.5

julia> distance(line, @Vec[1.0, 0.0], 0.5) === nothing
true

GeometricObjects.distance — Method

distance(::Sphere, x::Vec)
distance(::Sphere, x::Vec, threshold::Real)

Compute the distance vector from x to perpendicular foot on surface of sphere. When threshold is given, check the contact between line and point x, and return nothing if contact is not detected.

julia> sphere = Sphere(Vec(1.0,1.0), 1.0)
Sphere{2, Float64}:
  Coordinates: [[1.0, 1.0]]
  Attitude: [1.0, 0.0]

julia> d = distance(sphere, Vec(0.0,0.0))
2-element Vec{2, Float64}:
 0.29289321881345254
 0.29289321881345254

julia> norm(d) ≈ √2 - 1
true

GeometricObjects.perpendicularfoot — Method

GeometricObjects.perpendicularfoot(::Line, x::Vec)

Compute the position of perpendicular foot.